convert to float32 array

+++ lib.ts
@@ -190,8 +190,8 @@ export interface runOptions {
 }
 
 export const run = (program: Op[], opt?: runOptions) => {
-	const stack = new Uint16Array(STACK_SIZE)
-	const bus = new Uint16Array(BUS_SIZE)
+	const stack = new Float32Array(STACK_SIZE)
+	const bus = new Float32Array(BUS_SIZE)
 
 	let sp = 0
 	let fp = 0 // used to track the end of a stack independent of the stack pointer
